requirements.txt
hydra-core==1.0.6
numpy==1.20.3
pytorch-lightning==1.3.2
scikit-learn==0.24.2
scipy==1.6.3
# torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
transformers==4.6.1